# internal/controlplane/openapi.yaml
...
openapi: 3.1.0
info:
  title: FoxByte API
  version: "0.6"
  description: >
    Control-plane and Agent Branch APIs for FoxByte — serverless PostgreSQL
    with instant branching, time-travel, and Blackbox — a tamper-evident record of
    every schema change. Every `/api/branches/{name}/ledger…` path is also served
    at `/api/branches/{name}/blackbox…`.
    All endpoints are served over TLS. Authenticate with an API key
    (`key_…`) as a Bearer token.
  license:
    name: AGPL-3.0-or-later
servers:
  - url: https://localhost:8080
    description: Control-plane API
  - url: https://localhost:8088
    description: Agent Branch API
security:
  - apiKey: []
...
paths:
...
  /agents:
    get:
      tags: [agents]
      summary: List active agent branches. (Agent Branch API, :8088)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { agents: { type: array, items: { $ref: "#/components/schemas/AgentBranch" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /agents/{id}/branch:
    post:
      tags: [agents]
      summary: Create an agent's branch and return its DSN. (Agent Branch API, :8088)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: false
        description: >-
          Optional Blackbox provenance. Without it the request and response are unchanged. With it,
          a session is recorded on the agent's branch and every change the agent makes there carries
          the session, task and parent session; the response adds session_id (and the ids sent).
        content:
          application/json:
            schema:
              type: object
              properties:
                task_id: { type: string, maxLength: 200 }
                parent_session_id: { type: string, maxLength: 200 }
                session_id: { type: string, maxLength: 200, description: "Your own session id (default: generated)" }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/AgentBranch" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [agents]
      summary: Delete an agent's branch. (Agent Branch API, :8088)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: Deleted }, "401": { $ref: "#/components/responses/Unauthorized" } }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: An API key (`key_…`) minted with `fox apikey create`.
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid API key.
  schemas:
...
    AgentBranch:
      type: object
      properties:
        agent: { type: string }
        branch: { type: string }
        host: { type: string, description: "The Gateway's host: an agent connects through it, not to the branch's container." }
        port: { type: string, description: "The Gateway's port (6432 by default)." }
        dsn:
          type: string
          description: >-
            Connection string for this branch, routed through the Gateway, so it
            is reachable from the host and not only from inside the VM. The
            password is an API key scoped to this one branch: it opens no other
            branch, and the control plane and this API both refuse it. Returned
            once, at creation — listings omit the key — and revoked when the
            branch is deleted.
        status: { type: string, description: "ready, or suspended when the branch's container is stopped (the Gateway wakes it on connect)." }
...